requirements.txt
asgiref==3.7.2
astroid==3.3.9
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
defusedxml==0.7.1
dill==0.4.0
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.6
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
urllib3==2.4.0
drf-extra-fields==3.7.0
pytest==8.3.5
pytest-django==4.11.1
flake8==7.2.0
django-filter==25.1